backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
opencv-python==4.6.0.66
numpy==1.26.4
Pillow==10.1.0
deepface==0.0.87
tensorflow==2.15.0
firebase-admin==7.1.0
python-dotenv==1.1.1
websockets==15.0.1